pyproject.toml

[project]
name = "frame"
version = "0.1.0"
description = "A framework for machine learning tasks."
requires-python = ">=3.10,<3.14"
dependencies = [
    "hydra-core>=1.3.2",    
    "mlcroissant>=1.0.22",
    "mlflow>=3.6.0",
    "pytorch-lightning>=2.5.6",
    "ruff>=0.14.4",
    "submitit>=1.5.3",
    "torchgeo>=0.6.2",
    "torchx",
]

[tool.ruff]
line-length = 88

[tool.uv.sources]
torchx = { git = "https://github.com/meta-pytorch/torchx.git" }
